+++ packages/ricos-editor/web/src/utils/editorUtils.ts
@@ -1,5 +1,5 @@
 import { createEmpty, convertToRaw } from 'wix-rich-content-editor/dist/lib/editorStateConversion';
-import { EditorState } from 'draft-js';
+import { EditorState, ContentState } from 'draft-js';
 import { debounce } from 'lodash';
 import { emptyState } from 'ricos-common';
 import { isSSR } from 'wix-rich-content-common';
@@ -8,17 +8,22 @@ import { RicosContent, EditorDataInstance, OnContentChangeFunction } from '../in
 /* eslint-disable no-console */
 export const assert = (predicate, message) => console.assert(predicate, message);
 
-export function createDataConverter(lateOnChange?: OnContentChangeFunction): EditorDataInstance {
-  let currState: RicosContent = emptyState;
+export function createDataConverter(onContentChange?: OnContentChangeFunction): EditorDataInstance {
+  let currContent: RicosContent = emptyState;
   let currEditorState: EditorState = createEmpty();
+  let prevState: ContentState = currEditorState.getCurrentContent();
   let isUpdated = false;
   const getContentState = () => {
+    const currState: ContentState = currEditorState.getCurrentContent();
     if (!isUpdated) {
-      currState = convertToRaw(currEditorState.getCurrentContent());
+      currContent = convertToRaw(currState);
       isUpdated = true;
     }
-    lateOnChange?.(currState);
-    return currState;
+    if (currState !== prevState) {
+      onContentChange?.(currContent);
+      prevState = currState;
+    }
+    return currContent;
   };
   const debounceUpdate = debounce(getContentState, 200);
   return {